+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/auth/AuthRemoteDataSource.kt
@@ -1,12 +1,11 @@
-package com.app.edonymyeon.data.datasource.user
+package com.app.edonymyeon.data.datasource.auth
 
 import com.app.edonymyeon.data.dto.LoginDataModel
 import com.app.edonymyeon.data.service.AuthService
 import com.app.edonymyeon.data.service.client.RetrofitClient
 import retrofit2.Response
 
-class UserRemoteDataSource : UserDataSource {
-
+class AuthRemoteDataSource : AuthDataSource.Remote {
     private val authService: AuthService =
         RetrofitClient.getInstance().create(AuthService::class.java)
