+++ android/app/src/main/java/com/app/edonymyeon/data/repository/AuthRepositoryImpl.kt
@@ -2,19 +2,18 @@ package com.app.edonymyeon.data.repository
 
 import com.app.edonymyeon.data.common.CustomThrowable
 import com.app.edonymyeon.data.datasource.auth.AuthDataSource
-import com.app.edonymyeon.data.datasource.user.UserDataSource
 import com.app.edonymyeon.data.dto.LoginDataModel
 import com.domain.edonymyeon.repository.AuthRepository
 import org.json.JSONObject
 
 class AuthRepositoryImpl(
-    private val userDataSource: UserDataSource,
-    private val authDataSource: AuthDataSource,
+    private val authRemoteDataSource: AuthDataSource.Remote,
+    private val authLocalDataSource: AuthDataSource.Local,
 ) : AuthRepository {
     override suspend fun login(email: String, password: String): Result<Unit> {
-        val result = userDataSource.login(LoginDataModel(email, password))
+        val result = authRemoteDataSource.login(LoginDataModel(email, password))
         return if (result.isSuccessful) {
-            authDataSource.setAuthToken(result.headers()["Authorization"] as String)
+            authLocalDataSource.setAuthToken(result.headers()["Authorization"] as String)
             Result.success(result.body() ?: Unit)
         } else {
             val errorResponse = result.errorBody()?.string()
